Size Detection Mitigations / Various Refactoring

+++ windirstat/CsvLoader.cpp
@@ -226,7 +226,7 @@ bool SaveResults(const std::wstring& path, CItem * rootItem)
 
     // Vector to store all entries
     std::vector<const CItem*> items;
-    items.reserve(rootItem->GetItemsCount());
+    items.reserve(static_cast<size_t>(rootItem->GetItemsCount()));
 
     // Output all items to file
     std::stack<CItem*> queue({ rootItem });
